requirements.txt
streamlit
torch
torchvision
pillow
numpy
matplotlib
opencv-python 